TokenFusion/requirements.txt
# TokenFusion UAVScenes Requirements
# VM: CUDA 11.8, Python 3.10, Debian 11

# PyTorch (CUDA 11.8) - Install via pip with cu118
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.2+cu118
torchvision==0.16.2+cu118

# Vision Transformer utilities
timm>=0.9.0

# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
scipy>=1.11.0
tqdm>=4.65.0
Pillow>=10.0.0

# Configuration and logging
PyYAML>=6.0
tensorboard>=2.15.0

# Optional but recommended
matplotlib>=3.7.0
tabulate>=0.9.0
